_archive_/requirements.txt
boto3
# celery
# celery-redbeat
decorator
dj-database-url
dj-static
Django==4.0.0
# django-anymail[sendinblue]==7.1.0
# django-cache-memoize
# django-cors-headers
django-debug-toolbar
django-extensions
django-filter
django-redis-cache
# django-rest-auth
django-rest-swagger
# django-request
django-storages
django-timezone-field
# django-user-agents
django-widget-tweaks
djangorestframework
# djangorestframework-simplejwt
graphviz
ipython
ipywidgets
# markdownify
# matplotlib
msgpack
msgpack-numpy
Pillow
plotly
popoto>=0.7.3
python-telegram-bot
pytz
PyYAML
redis==3.5.3
redisgraph
# requests-oauthlib==1.3.0
requests
schedule
simplejson
six
toolz
waitress

# FINANCE
# backtrader
# yahoo_fin
# alpha_vantage # for TA indicators
# quandl # for frequent price calls
# polygon-api-client  # for frequent price calls

# AI LIBRARIES
numpy
pandas
streamlit
# openai

# LOCAL DEV
# psycopg2-binary

# INSTALLED VIA ANACONDA ON LOCAL (in order):
# -c apple tensorflow-deps
# opencv
# scikit_learn
# scikit-image
# keras
# openblas
# scipy
# torch
# torchvision
# fastai
# fastai2
# fastcore


# PRODUCTION ONLY
# psycopg2
# https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
# tensorflow-cpu


# OTHER
# azure-cognitiveservices-search-imagesearch
# sentencepiece
